Integration Tests:   - Refactor resource loading   - Move tests into integration test module
