+++ test/integration/elastic_search/docker-compose.yml
