+++ test/integration/elastic_search/keycloak-init/dataportal-realm.json
